test: check resource usage data in telemetry test
