+++ misc/telemetry/Dockerfile
@@ -10,5 +10,13 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
+FROM golang:1.7
 
-docker build -t "amazon/amazon-ecs-windows-cpupercent-test:make" -f "${PSScriptRoot}/windows.dockerfile" ${PSScriptRoot}
+WORKDIR /gopath
+
+COPY main.go .
+
+RUN go build -o stress main.go
+
+ENTRYPOINT ["./stress"]
+CMD [ "-concurrency", "1000", "-memory", "1024"]
