+++ misc/telemetry/Makefile
@@ -1,0 +1,6 @@
+.PHONY: all
+all:
+	docker build -t amazon/amazon-ecs-telemetry-test:make .
+
+clean:
+	-docker rmi -f "amazon/amazon-ecs-telemetry-test:make"
